+++ NEWS.md
@@ -1,6 +1,7 @@
 # aqp 1.30 (2021-04-14)
  * `mixMunsell`:
     - new method `exact` for direct conversion of mixture spectra to sRGB or closest Munsell chip (via `spec2Munsell()`)
+ * new convenience function `PMS2Munsell()` for converting PMS codes -> closest Munsell chip (https://github.com/ncss-tech/aqp/issues/124)
 
 # aqp 1.29 (2021-04-05)
  * CRAN release